react/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cmake==3.25.0
colorama==0.4.6
cryptography==43.0.3
dataclasses-json==0.6.7
dlib==19.24.2
dnspython==2.7.0
docarray==0.40.0
docx2txt==0.8
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.0
firebase-admin==6.5.0
frozenlist==1.5.0
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-ollama==0.2.0
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.142
lxml==5.3.0
markdown-it-py==3.0.0
marshmallow==3.23.1
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
ollama==0.3.3
opencv-python==4.10.0.84
orjson==3.10.11
packaging==24.2
passlib==1.7.4
pillow==11.0.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.10.1
pyodbc==5.1.0
pyparsing==3.2.0
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlmodel==0.0.22
starlette==0.38.6
tenacity==9.0.0
threadpoolctl==3.5.0
tqdm==4.67.0
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.0
yarl==1.17.1
